requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
